Combine EntityScene to Scene
